pyproject.toml
[project]
name = "market-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "passlib[argon2]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
